requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
beautifulsoup4~=4.12.3
blinker
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
cssselect2
Cython
dataclasses-json
decorator
distro
dtw-python
edge-tts
exceptiongroup
filelock
fonttools
frozenlist
fsspec
gitdb
GitPython
greenlet
h11
html5lib
httpcore
httpx
idna
imageio
imageio-ffmpeg
importlib_metadata
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain~=0.2.16
langchain-community~=0.2.16
langchain-core~=0.2.38
langchain-openai~=0.1.23
langchain-text-splitters~=0.2.4
langchainhub
langgraph~=0.2.16
langsmith
llvmlite
Markdown
markdown-it-py
MarkupSafe
marshmallow
mdurl
more-itertools
moviepy
mpmath
multidict
mypy-extensions
narwhals
networkx
numba
numexpr~=2.10.1
numpy
openai~=1.43.0
openai-whisper
orjson
packaging
pandas
pdfkit
pillow~=10.4.0
pip
proglog
protobuf
pyarrow
pycparser
pydantic~=2.8.2
pydantic_core
pydeck
pydyf
Pygments
pyphen
PySocks
python-dateutil
python-decouple~=3.8
pytz
PyYAML
referencing
regex
reportlab~=4.2.2
requests~=2.32.3
rich
rpds-py
scipy
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
streamlit~=1.38.0
sympy
tenacity
tiktoken
tinycss2
toml
torch
tornado
tqdm
types-requests
typing_extensions
typing-inspect
tzdata
urllib3
watchdog
weasyprint
webencodings
wheel
whisper-timestamped
win-inet-pton
yarl
zipp
zopfli

langchain-anthropic~=0.1.23
gTTS~=2.5.3
boto3~=1.35.21
mysql~=0.0.3
botocore~=1.35.21